backend/requirements.txt
requests
numpy
python-dotenv  # if using .env for keys
